+++ wgpu-core/src/id.rs
@@ -4,7 +4,6 @@ use core::{
     fmt::{self, Debug},
     hash::Hash,
     marker::PhantomData,
-    mem::size_of,
     num::NonZeroU64,
 };
 use wgt::WasmNotSendSync;