+++ wgpu-core/src/registry.rs
@@ -1,5 +1,4 @@
 use alloc::sync::Arc;
-use core::mem::size_of;
 
 use crate::{
     id::Id,