+++ wgpu-hal/examples/ray-traced-triangle/main.rs
@@ -8,9 +8,7 @@ use raw_window_handle::{HasDisplayHandle, HasWindowHandle};
 use glam::{Affine3A, Mat4, Vec3};
 use std::{
     borrow::{Borrow, Cow},
-    iter,
-    mem::size_of,
-    ptr,
+    iter, ptr,
     time::Instant,
 };
 use wgpu_types::Dx12BackendOptions;