+++ wgpu-hal/src/dx12/adapter.rs
@@ -1,11 +1,4 @@
-use std::{
-    mem::{size_of, size_of_val},
-    ptr,
-    string::String,
-    sync::Arc,
-    thread,
-    vec::Vec,
-};
+use std::{ptr, string::String, sync::Arc, thread, vec::Vec};
 
 use parking_lot::Mutex;
 use windows::{